moves cubemap to util
